+++ src/pages/shoppingCart.style.ts
@@ -1,0 +1,19 @@
+import { css } from "@emotion/react";
+
+const titleStyle = css`
+  font-weight: 700;
+  font-size: 24px;
+`;
+
+const subTitleStyle = css`
+  font-weight: 500;
+  font-size: 12px;
+`;
+
+const titleBox = css`
+  display: flex;
+  gap: 15px;
+  flex-direction: column;
+`;
+
+export { titleStyle, subTitleStyle, titleBox };
